fix: builds without typesafe i18n keys
