+++ packages/sanity/src/core/config/studio/types.ts
@@ -1,6 +1,5 @@
-import type {InitOptions, ResourceLanguage} from 'i18next'
 import React from 'react'
-import {ConfigContext, LanguageDefinition, LanguageLoader, Tool} from '../types'
+import {Tool} from '../types'
 
 /**
  * @hidden
@@ -56,103 +55,4 @@ export interface StudioComponentsPluginOptions {
   logo?: React.ComponentType<LogoProps>
   navbar?: React.ComponentType<NavbarProps>
   toolMenu?: React.ComponentType<ToolMenuProps>
-}
-
-/** @alpha */
-export interface I18nContext {
-  projectId: string
-  dataset: string
-}
-
-/** @alpha */
-export interface LanguageBundle {
-  namespace: string
-  resources: ResourceLanguage
-  /** Should the resources be merged deeply (nested objects). Default: true */
-  deep?: boolean
-  /** Should existing resource keys for the namespace be overwritten. Default: false */
-  overwrite?: boolean
-}
-
-/** @alpha */
-export type I18nLanguagesOption =
-  | ((prev: LanguageDefinition[], context: I18nContext) => LanguageDefinition[])
-  | LanguageDefinition[]
-
-/** @alpha */
-export type I18nLanguageLoaderOption =
-  | ((prev: LanguageLoader[], context: I18nContext) => LanguageLoader[])
-  | LanguageLoader[]
-
-/** @alpha */
-export interface I18nPluginOptions {
-  /**
-   * Defines which languages should be available for user selection.
-   * Prev is initially `[{id: 'en-US', title: 'English (US)', icon: AmericanFlag }]`
-   *
-   * Language titles and icons can be changed by transforming the LanguageDefinition array values.
-   *
-   * User selected language
-   */
-  languages?: I18nLanguagesOption
-
-  /**
-   * Allows redefining the I18next init options before they are used.
-   * Invoked when a workspace is loaded
-   */
-  initOptions?: (options: InitOptions, context: I18nContext) => InitOptions
-
-  /**
-   * Defines language bundles that will be loaded lazily.
-   *
-   * ### Example
-   *
-   * ```ts
-   *
-   * ```
-   *
-   */
-  languageLoaders?: I18nLanguageLoaderOption
-
-  /**
-   * When this is true, schema type title and descriptions will be translated.
-   * Configure a languageLoader that returns a language bundle for the `schema` namespace,
-   * with resource keys using the following convention:
-   *
-   * ## Keys for types
-   * - `<typeName>|title`
-   * - `<typeName>|description`
-   *
-   * ## Keys for fields
-   *
-   * -`<objectTypeName>.<fieldName>|title`
-   * -`<objectTypeName>.<fieldName>|description`
-   *
-   * ## Keys for array items
-   *
-   * - `<arrayTypeName>.<arrayMemberTypeName>|title`
-   * - `<arrayTypeName>.<arrayMemberTypeName>|description`
-   *
-   * ## Caveats
-   *
-   * Enabling schema translations could adversely impact studio performance.
-   * Inline definitions for objects are not supported (nested types).
-   *
-   * ## Example LanguageBundle
-   *
-   *```ts
-   * // locales/en_US/schema.ts
-   * export default {
-   *   namespace: 'schema',
-   *   resources: {
-   *     'myDocumentType|title': 'Document type 'myDocumentType' will use this string as title wherever it is used',
-   *
-   *     'myDocumentType.text|title': 'Document field named 'text' will use this string as title'
-   *     'myDocumentType.text|description': 'Document field named 'text' will this string as description',
-   *   },
-   *  }
-   *```
-   *
-   */
-  experimentalTranslateSchemas?: boolean
 }
