+++ packages/sanity/src/core/config/types.ts
@@ -11,7 +11,7 @@ import type {
 } from '@sanity/types'
 import React, {type ComponentType} from 'react'
 import type {Observable} from 'rxjs'
-import type {InitOptions, i18n, TFunction} from 'i18next'
+import type {i18n, InitOptions, TFunction, ResourceLanguage} from 'i18next'
 import type {
   BlockAnnotationProps,
   BlockProps,
@@ -35,12 +35,7 @@ import {
   DocumentFieldActionsResolverContext,
   DocumentInspector,
 } from './document'
-import {
-  I18nPluginOptions,
-  LanguageBundle,
-  StudioComponents,
-  StudioComponentsPluginOptions,
-} from './studio'
+import {StudioComponents, StudioComponentsPluginOptions} from './studio'
 import {Router, RouterState} from 'sanity/router'
 
 /**
@@ -268,7 +263,121 @@ export type DocumentInspectorsResolver = ComposableOption<
 
 /**
  * @hidden
- * @beta */
+ * @beta
+ */
+export interface I18nContext {
+  projectId: string
+  dataset: string
+}
+
+/**
+ * @hidden
+ * @beta
+ */
+export interface LanguageBundle {
+  namespace: string
+  resources: ResourceLanguage
+  /** Should the resources be merged deeply (nested objects). Default: true */
+  deep?: boolean
+  /** Should existing resource keys for the namespace be overwritten. Default: false */
+  overwrite?: boolean
+}
+
+/** @beta @hidden */
+export type I18nLanguagesOption =
+  | ((prev: LanguageDefinition[], context: I18nContext) => LanguageDefinition[])
+  | LanguageDefinition[]
+
+/** @beta @hidden */
+export type I18nLanguageLoaderOption =
+  | ((prev: LanguageLoader[], context: I18nContext) => LanguageLoader[])
+  | LanguageLoader[]
+
+/** @beta @hidden */
+export interface I18nPluginOptions {
+  /**
+   * Defines which languages should be available for user selection.
+   * Prev is initially `[{id: 'en-US', title: 'English (US)', icon: AmericanFlag }]`
+   *
+   * Language titles and icons can be changed by transforming the LanguageDefinition array values.
+   *
+   * User selected language
+   */
+  languages?: I18nLanguagesOption
+
+  /**
+   * Allows redefining the I18next init options before they are used.
+   * Invoked when a workspace is loaded
+   */
+  initOptions?: (options: InitOptions, context: I18nContext) => InitOptions
+
+  /**
+   * Defines language bundles that will be loaded lazily.
+   *
+   * ### Example
+   *
+   * ```ts
+   *
+   * ```
+   *
+   */
+  languageLoaders?: I18nLanguageLoaderOption
+
+  /**
+   * When this is true, schema type title and descriptions will be translated.
+   * Configure a languageLoader that returns a language bundle for the `schema` namespace,
+   * with resource keys using the following convention:
+   *
+   * ## Keys for types
+   * - `<typeName>|title`
+   * - `<typeName>|description`
+   *
+   * ## Keys for fields
+   *
+   * -`<objectTypeName>.<fieldName>|title`
+   * -`<objectTypeName>.<fieldName>|description`
+   *
+   * ## Keys for array items
+   *
+   * - `<arrayTypeName>.<arrayMemberTypeName>|title`
+   * - `<arrayTypeName>.<arrayMemberTypeName>|description`
+   *
+   * ## Caveats
+   *
+   * Enabling schema translations could adversely impact studio performance.
+   * Inline definitions for objects are not supported (nested types).
+   *
+   * ## Example LanguageBundle
+   *
+   *```ts
+   * // locales/en_US/schema.ts
+   * export default {
+   *   namespace: 'schema',
+   *   resources: {
+   *     'myDocumentType|title': 'Document type 'myDocumentType' will use this string as title wherever it is used',
+   *
+   *     'myDocumentType.text|title': 'Document field named 'text' will use this string as title'
+   *     'myDocumentType.text|description': 'Document field named 'text' will this string as description',
+   *   },
+   *  }
+   *```
+   *
+   * Now, in studio.config.ts:
+   * ```ts
+   * defineConfig({
+   *   i18n: {
+   *     experimentalTranslateSchemas: true,
+   *     languageLoaders: [
+   *         (lang) => import(`./locales/${lang}/schema.ts`),
+   *     ]
+   *   },
+   * })
+   * ```
+   */
+  experimentalTranslateSchemas?: boolean
+}
+
+/** @beta */
 export interface PluginOptions {
   name: string
   plugins?: PluginOptions[]
@@ -281,7 +390,7 @@ export interface PluginOptions {
   studio?: {
     components?: StudioComponentsPluginOptions
   }
-  /** alpha */
+  /** @beta @hidden */
   i18n?: I18nPluginOptions
 }
 
@@ -447,7 +556,7 @@ export interface Source {
      * @beta */
     resolveNewDocumentOptions: (context: NewDocumentCreationContext) => InitialValueTemplateItem[]
 
-    /** @alpha */
+    /** @beta @hidden */
     unstable_languageFilter: (
       props: PartialContext<DocumentLanguageFilterContext>
     ) => DocumentLanguageFilterComponent[]
@@ -502,14 +611,14 @@ export interface Source {
     components?: StudioComponents
   }
 
-  /** @alpha */
-  i18n: I18nSource
-
-  /** @alpha */
+  /** @beta @hidden */
   search: {
     filters: SearchFilterDefinition[]
     operators: SearchOperatorDefinition[]
   }
+
+  /** @internal */
+  i18n: I18nSource
 
   /** @internal */
   __internal: {
@@ -519,15 +628,17 @@ export interface Source {
   }
 }
 
+/** @beta @hidden */
 export interface LanguageDefinition {
   id: string
   title: string
   icon?: ComponentType
 }
 
+/** @beta @hidden */
 export type LanguageLoaderResult = Promise<LanguageBundle | {default: LanguageBundle} | undefined>
 
-// TODO return observable instead of Promise?
+/** @beta @hidden */
 export type LanguageLoader = (language: string, context: {i18n: i18n}) => LanguageLoaderResult
 
 /** @internal */
