+++ packages/sanity/src/core/i18n/i18nConfig.ts
@@ -1,5 +1,5 @@
-import i18nApi, {i18n, InitOptions, Resource} from 'i18next'
-import {Schema, typed} from '@sanity/types'
+import i18nApi, {type i18n, type InitOptions} from 'i18next'
+import {Schema} from '@sanity/types'
 import {I18nSource, LanguageDefinition, LanguageLoader, SourceOptions} from '../config'
 import {resolveConfigProperty} from '../config/resolveConfigProperty'
 import {
@@ -13,7 +13,6 @@ import {getPreferredLang} from './languageStore'
 import {schemaI18nNamespace, studioI18nNamespace} from './i18nNamespaces'
 import {studioLocaleLoader} from './studioLocaleLoader'
 import {i18nSchema} from './i18nSchema'
-import {SanityLanguageResources} from './locales/types'
 
 export const defaultI18nOptions: InitOptions = {
   partialBundledLanguages: true,
@@ -21,10 +20,10 @@ export const defaultI18nOptions: InitOptions = {
   lng: defaultLanguage.id,
   fallbackLng: defaultLanguage.id,
   resources: {
-    [defaultLanguage.id]: typed<Partial<SanityLanguageResources>>({
+    [defaultLanguage.id]: {
       [studioI18nNamespace]: studioI18nNamespaceStrings,
       [schemaI18nNamespace]: {},
-    }),
+    },
   },
   debug: false,
   initImmediate: false,
