+++ packages/sanity/src/core/i18n/i18nNamespaces.ts
@@ -1,4 +1,8 @@
+/** @alpha */
 export const studioI18nNamespace = 'studio' as const
 
-/** By-convention namespace for schema translations. Use to just-in-time translate schema title and description. */
+/**
+ * By-convention namespace for schema translations. Use to just-in-time translate schema title and description.
+ * @alpha
+ * */
 export const schemaI18nNamespace = 'schema' as const
