+++ packages/sanity/src/core/i18n/locales/types/schema.ts
@@ -1,3 +1,6 @@
 // interface available for extension
+/**
+ * @alpha
+ */
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface SchemaTranslations {}
