+++ packages/sanity/src/core/i18n/locales/types/studio.ts
@@ -1,3 +1,6 @@
 import {studioI18nNamespaceStrings} from '../en-US/studio'
 
+/**
+ * @alpha
+ */
 export type StudioTranslations = typeof studioI18nNamespaceStrings
