+++ packages/sanity/src/core/i18n/localizedLanguages.tsx
@@ -2,6 +2,10 @@ import React from 'react'
 import {typed} from '@sanity/types'
 import {LanguageDefinition} from '../config'
 
+/**
+ * Languages that the studio has available LanguageBundles for
+ * @alpha
+ */
 export const localizedLanguages = {
   'en-US': typed<LanguageDefinition>({
     id: 'en-US',
@@ -17,4 +21,7 @@ export const localizedLanguages = {
   }),
 }
 
+/**
+ * @alpha
+ */
 export const defaultLanguage: LanguageDefinition = localizedLanguages['en-US']
