+++ packages/sanity/src/core/studio/components/navbar/search/components/PlaceholderSearchInput.tsx
@@ -2,10 +2,10 @@
 import {SearchIcon} from '@sanity/icons'
 import {Flex, KBD, TextInput} from '@sanity/ui'
 import React, {forwardRef, KeyboardEvent as ReactKeyboardEvent, Ref, useCallback} from 'react'
+import {useTranslation} from 'react-i18next'
 import styled from 'styled-components'
 import {GLOBAL_SEARCH_KEY, GLOBAL_SEARCH_KEY_MODIFIER} from '../constants'
 import {useSearchState} from '../contexts/search/useSearchState'
-import {useTranslation} from '../../../../../i18n'
 
 const KeyboardShortcutFlex = styled(Flex)`
   position: absolute;
