+++ packages/sanity/src/desk/deskTool.ts
@@ -13,7 +13,7 @@ import {getIntentState} from './getIntentState'
 import {router} from './router'
 import {DeskToolOptions} from './types'
 import {changesInspector} from './panes/document/inspectors/changes'
-import {deskI18nNamespace} from './i18n'
+import {deskI18nNamespace} from './i18n/i18nNamespaces'
 import {deskI18nNamespaceStrings} from './i18n/locales/en-US/desk'
 import {deskLocaleLoader} from './i18n/deskLocaleLoader'
 import {validationInspector} from './panes/document/inspectors/validation'
