+++ packages/sanity/src/desk/i18n/i18nNamespaces.ts
@@ -1,1 +1,4 @@
+/**
+ * @alpha
+ */
 export const deskI18nNamespace = 'desk' as const
