+++ packages/sanity/src/desk/i18n/index.ts
@@ -1,2 +1,2 @@
 export * from './i18nNamespaces'
-export {type DeskTranslations} from './locales/types/desk'
+export * from './locales/types/desk'
